Monkey patch existing controller class instead of defining a new controller
We do so by adding a callback - to be executed before each request in development, and at startup in production - to patch the existing app classes. Doing so in init/environment.rb wouldn't work in development, since classes are reloaded but initialization is not run each time. See http://stackoverflow.com/questions/7072758/plugin-not-reloading-in-development-mode

Creating a new controller was a workaround since I couldn't get the monkey patch to work before: as mentioned, classes are reloaded, but plugins are not, so just patching in the theme won't work; plus there are strange interactions with the order classes are loaded which required additional patches like https://github.com/sebbacon/alaveteli/commit/992a0b8acc4556fa1687bdf7bc66d63a51d410b3 Plus sometimes in development it resulted in strange "A copy of ApplicationController has been removed from the module tree but is still active!" error messages.